pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["alfarizi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.7"
pyjwt = "^2.9.0"
flask-bcrypt = "^1.0.1"
flask-swagger-ui = "^4.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
